extract DLProxy interface
