+++ DynamicLoadApk/lib/src/com/ryg/dynamicload/DLProxyFragmentActivity.java
@@ -31,8 +31,8 @@ import android.view.MotionEvent;
 import android.view.WindowManager.LayoutParams;
 
 import com.ryg.dynamicload.internal.DLPluginManager;
+import com.ryg.dynamicload.internal.DLProxy;
 import com.ryg.dynamicload.internal.DLProxyImpl;
-import com.ryg.dynamicload.internal.DLProxyImpl.DLProxy;
 
 public class DLProxyFragmentActivity extends FragmentActivity implements DLProxy {
